==> 06_further_analyisis.ipynb <==
import pandas as pd
from matplotlib.colors import ListedColormap
from faerun import Faerun
from rdkit import Chem 
from rdkit.Chem import Lipinski, Descriptors, rdMolDescriptors, AllChem, PandasTools, rdchem
from pandarallel import pandarallel
import numpy as np
pandarallel.initialize(progress_bar=False)
import tmap as tm
from map4 import MAP4Calculator
import os
import seaborn as sns
import joblib
import matplotlib.pyplot as plt
from rdkit.Chem.AtomPairs import Pairs
folder = "/data/coconut/"

coconut = pd.read_pickle(folder + "coconut_prop.pkl")

coconut.origin.unique()

extra = coconut.query("simple_tax=='other'").copy()

coconut = coconut.query("origin == 0 or origin == 1 or origin == 2").copy()
coconut.replace(["NaN", 'NaT'], np.nan, inplace = True)
coconut = coconut.dropna()

len(coconut)

np.max(np.array(coconut.MW.to_list()))

np.min(np.array(coconut.MW.to_list()))

coconut["LOW_fcsp3"] = coconut.fcsp3.map(lambda x: x<=0.2)
coconut["MED_fcsp3"] = coconut.fcsp3.map(lambda x: x<0.8)
coconut["HIGH_fcsp3"] = coconut.fcsp3.map(lambda x: x>=0.8)
coconut["LOW_mw"] = coconut.MW.map(lambda x: x<=300)
coconut["MED_mw"] = coconut.MW.map(lambda x: x<800)
coconut["HIGH_mw"] = coconut.MW.map(lambda x: x>=800)

coconut.alogp

coconut["LOW_alogp"] = coconut.alogp.map(lambda x: float(x)<=-2)
coconut["MED_alogp"] = coconut.alogp.map(lambda x: float(x)<8)
coconut["HIGH_alogp"] = coconut.alogp.map(lambda x: float(x)>=8)

LOW_alogp = len(coconut.query("LOW_alogp==True"))
print("LOW_alogp", LOW_alogp)
LOW_fcsp3_plants = len(coconut.query("LOW_alogp==True and origin==0"))
print("LOW_alogp_plants", LOW_fcsp3_plants)
LOW_alogp_fungi = len(coconut.query("LOW_alogp==True and origin==1"))
print("LOW_alogp_fungi", LOW_alogp_fungi)
LOW_alogp_bacteria = len(coconut.query("LOW_alogp==True and origin==2"))
print("LOW_alogp_bacteria", LOW_alogp_bacteria)

MED_alogp = len(coconut.query("MED_alogp==True and LOW_fcsp3==False"))
print("MED_alogp", MED_alogp)
MED_alogp_plant = len(coconut.query("MED_alogp==True and LOW_alogp==False and origin==0"))
print("MED_alogp_plant", MED_alogp_plant)
MED_alogp_fungi = len(coconut.query("MED_alogp==True and LOW_alogp==False and origin==1"))
print("MED_alogp_fungi", MED_alogp_fungi)
MED_alogp_bacteria = len(coconut.query("MED_alogp==True and LOW_alogp==False and origin==2"))
print("MED_alogp_bacteria", MED_alogp_bacteria)

HIGH_alogp = len(coconut.query("HIGH_alogp==True"))
print("HIGH_alogp", HIGH_alogp)
HIGH_alogp_plants = len(coconut.query("HIGH_alogp==True and origin==0"))
print("HIGH_alogp_plants", HIGH_alogp_plants)
HIGH_alogp_fungi = len(coconut.query("HIGH_alogp==True and origin==1"))
print("HIGH_alogp_fungi", HIGH_alogp_fungi)
HIGH_alogp_bacteria = len(coconut.query("HIGH_alogp==True and origin==2"))
print("HIGH_alogp_bacteria", HIGH_alogp_bacteria)

LOW_fcsp3 = len(coconut.query("LOW_fcsp3==True"))
print("LOW_fcsp3", LOW_fcsp3)
LOW_fcsp3_plants = len(coconut.query("LOW_fcsp3==True and origin==0"))
print("LOW_fcsp3_plants", LOW_fcsp3_plants)
LOW_fcsp3_fungi = len(coconut.query("LOW_fcsp3==True and origin==1"))
print("LOW_fcsp3_fungi", LOW_fcsp3_fungi)
LOW_fcsp3_bacteria = len(coconut.query("LOW_fcsp3==True and origin==2"))
print("LOW_fcsp3_bacteria", LOW_fcsp3_bacteria)

MED_fcsp3 = len(coconut.query("MED_fcsp3==True and LOW_fcsp3==False"))
print("MED_fcsp3", MED_fcsp3)
MED_fcsp3_plant = len(coconut.query("MED_fcsp3==True and LOW_fcsp3==False and origin==0"))
print("MED_fcsp3_plant", MED_fcsp3_plant)
MED_fcsp3_fungi = len(coconut.query("MED_fcsp3==True and LOW_fcsp3==False and origin==1"))
print("MED_fcsp3_fungi", MED_fcsp3_fungi)
MED_fcsp3_bacteria = len(coconut.query("MED_fcsp3==True and LOW_fcsp3==False and origin==2"))
print("MED_fcsp3_bacteria", MED_fcsp3_bacteria)

HIGH_fcsp3 = len(coconut.query("HIGH_fcsp3==True"))
print("HIGH_fcsp3", HIGH_fcsp3)
HIGH_fcsp3_plants = len(coconut.query("HIGH_fcsp3==True and origin==0"))
print("HIGH_fcsp3_plants", HIGH_fcsp3_plants)
HIGH_fcsp3_fungi = len(coconut.query("HIGH_fcsp3==True and origin==1"))
print("HIGH_fcsp3_fungi", HIGH_fcsp3_fungi)
HIGH_fcsp3_bacteria = len(coconut.query("HIGH_fcsp3==True and origin==2"))
print("HIGH_fcsp3_bacteria", HIGH_fcsp3_bacteria)

HIGH_fcsp3+MED_fcsp3+LOW_fcsp3

LOW_mw= len(coconut.query("LOW_mw==True"))
print("LOW_mw", LOW_mw)
LOW_mw_plant = len(coconut.query("LOW_mw==True and origin ==0"))
print("LOW_mw_plant", LOW_mw_plant)
LOW_mw_fungi = len(coconut.query("LOW_mw==True and origin ==1"))
print("LOW_mw_fungi", LOW_mw_fungi)
LOW_mw_bacteria = len(coconut.query("LOW_mw==True and origin ==2"))
print("LOW_mw_bacteria", LOW_mw_bacteria)

HIGH_mw=len(coconut.query("HIGH_mw==True"))
print("HIGH_mw", HIGH_mw)
HIGH_mw_plant=len(coconut.query("HIGH_mw==True and origin==0"))
print("HIGH_mw_plant", HIGH_mw_plant)
HIGH_mw_fungi=len(coconut.query("HIGH_mw==True and origin==1"))
print("HIGH_mw_fungi", HIGH_mw_fungi)
HIGH_mw_bacteria=len(coconut.query("HIGH_mw==True and origin==2"))
print("HIGH_mw_bacteria", HIGH_mw_bacteria)

HIGH_mw_bacteria + HIGH_mw_fungi+ HIGH_mw_plant

MED_mw=len(coconut.query("MED_mw==True and LOW_mw==False"))
print("MED_mw", MED_mw)
MED_mw_plants=len(coconut.query("MED_mw==True and LOW_mw==False and origin==0"))
print("MED_mw_plants", MED_mw_plants)
MED_mw_fungi=len(coconut.query("MED_mw==True and LOW_mw==False and origin ==1"))
print("MED_mw_fungi", MED_mw_fungi)
MED_mw_bacteria=len(coconut.query("MED_mw==True and LOW_mw==False and origin==2"))
print("MED_mw_bacteria", MED_mw_bacteria)

MED_mw+HIGH_mw+LOW_mw

len(coconut)

peptide=len(coconut.query("isPeptide==True"))
print("peptide", peptide)
peptide_plant=len(coconut.query("isPeptide==True and origin==0"))
print("peptide_plant", peptide_plant)
peptide_fungi=len(coconut.query("isPeptide==True and origin==1"))
print("peptide_fungi", peptide_fungi)
peptide_bacteria=len(coconut.query("isPeptide==True and origin==2"))
print("peptide_bacteria", peptide_bacteria)

sugar=len(coconut.query("hasSugar==True"))
print("sugar", sugar)
sugar_plants=len(coconut.query("hasSugar==True and origin ==0"))
print("sugar_plants", sugar_plants)
sugar_fungi=len(coconut.query("hasSugar==True and origin ==1"))
print("sugar_fungi", sugar_fungi)
sugar_bacteria=len(coconut.query("hasSugar==True and origin ==2"))
print("sugar_bacteria", sugar_bacteria)

plant = len(coconut.query("origin==0"))
print("plant", plant)
fungi = len(coconut.query("origin==1"))
print("fungi", fungi)
bacteria = len(coconut.query("origin==2"))
print("bacteria", bacteria)

SVM = joblib.load(folder + f"MAP4-SVM-coconut.all.pkl")

MAP4 = MAP4Calculator(dimensions=1024)
def calc_map4(smiles):
    mol = Chem.MolFromSmiles(smiles)
    smiles = Chem.MolToSmiles(mol, isomericSmiles=False)
    mol = Chem.MolFromSmiles(smiles)
    map4 = MAP4.calculate(mol)
    return np.array(map4)

def map4_kernel_SVM(a, b=None):
    if b is None:
        b=a
    JS_all_pairs = np.zeros((len(a),len(b)))
    for i,fp1 in enumerate(a):
        for j,fp2 in enumerate(b):
            JS_all_pairs[i,j] = np.float(np.count_nonzero(fp1 == fp2)) / np.float(len(fp1))
    return JS_all_pairs

extra["map4"] = extra.SMILES.map(calc_map4)

extra["prediction"] = extra.map4.map(lambda x : SVM.predict_proba([x]))

extra["prediction_"] = extra.map4.map(lambda x : SVM.predict([x]))

def clean_pred(pred_):
    dict = {0:"plants", 1:"fungi", 2:"bacteria"}
    return dict[pred_[0]]

extra["predicted_organism"] = extra.prediction_.map(clean_pred)

extra_=extra[["coconut_id", "textTaxa", "SMILES", 'citationDOI', 'name', 'synonyms']].copy()
PandasTools.AddMoleculeColumnToFrame(extra_, "SMILES", "molecule", includeFingerprints=False)
PandasTools.SaveXlsxFromFrame(extra_, "other_taxa_with_prediction.xlsx", "molecule")

extra["LOW_fcsp3"] = extra.fcsp3.map(lambda x: x<=0.2)
extra["MED_fcsp3"] = extra.fcsp3.map(lambda x: x<0.8)
extra["HIGH_fcsp3"] = extra.fcsp3.map(lambda x: x>=0.8)
extra["LOW_mw"] = extra.MW.map(lambda x: x<=300)
extra["MED_mw"] = extra.MW.map(lambda x: x<800)
extra["HIGH_mw"] = extra.MW.map(lambda x: x>=800)

len(extra.query("predicted_organism == 'bacteria'"))

LOW_fcsp3 = len(extra.query("LOW_fcsp3==True"))
print("LOW_fcsp3", LOW_fcsp3)
LOW_fcsp3_plants = len(extra.query("LOW_fcsp3==True and predicted_organism == 'plants'"))
print("LOW_fcsp3_plants", LOW_fcsp3_plants)
LOW_fcsp3_fungi = len(extra.query("LOW_fcsp3==True and predicted_organism == 'fungi'"))
print("LOW_fcsp3_fungi", LOW_fcsp3_fungi)
LOW_fcsp3_bacteria = len(extra.query("LOW_fcsp3==True and predicted_organism == 'bacteria'"))
print("LOW_fcsp3_bacteria", LOW_fcsp3_bacteria)
print("******")

MED_fcsp3 = len(extra.query("MED_fcsp3==True and LOW_fcsp3==False"))
print("MED_fcsp3", MED_fcsp3)
MED_fcsp3_plant = len(extra.query("MED_fcsp3==True and LOW_fcsp3==False and predicted_organism == 'plants'"))
print("MED_fcsp3_plant", MED_fcsp3_plant)
MED_fcsp3_fungi = len(extra.query("MED_fcsp3==True and LOW_fcsp3==False and predicted_organism == 'fungi'"))
print("MED_fcsp3_fungi", MED_fcsp3_fungi)
MED_fcsp3_bacteria = len(extra.query("MED_fcsp3==True and LOW_fcsp3==False and predicted_organism == 'bacteria'"))
print("MED_fcsp3_bacteria", MED_fcsp3_bacteria)
print("******")

HIGH_fcsp3 = len(extra.query("HIGH_fcsp3==True"))
print("HIGH_fcsp3", HIGH_fcsp3)
HIGH_fcsp3_plants = len(extra.query("HIGH_fcsp3==True and predicted_organism == 'plants'"))
print("HIGH_fcsp3_plants", HIGH_fcsp3_plants)
HIGH_fcsp3_fungi = len(extra.query("HIGH_fcsp3==True and predicted_organism == 'fungi'"))
print("HIGH_fcsp3_fungi", HIGH_fcsp3_fungi)
HIGH_fcsp3_bacteria = len(extra.query("HIGH_fcsp3==True and predicted_organism == 'bacteria'"))
print("HIGH_fcsp3_bacteria", HIGH_fcsp3_bacteria)
print("******")

LOW_mw= len(extra.query("LOW_mw==True"))
print("LOW_mw", LOW_mw)
LOW_mw_plant = len(extra.query("LOW_mw==True and predicted_organism == 'plants'"))
print("LOW_mw_plant", LOW_mw_plant)
LOW_mw_fungi = len(extra.query("LOW_mw==True and predicted_organism == 'fungi'"))
print("LOW_mw_fungi", LOW_mw_fungi)
LOW_mw_bacteria = len(extra.query("LOW_mw==True and predicted_organism == 'bacteria'"))
print("LOW_mw_bacteria", LOW_mw_bacteria)
print(LOW_mw_bacteria+LOW_mw_fungi+LOW_mw_plant)
print("******")

HIGH_mw=len(extra.query("HIGH_mw==True"))
print("HIGH_mw", HIGH_mw)
HIGH_mw_plant=len(extra.query("HIGH_mw==True and predicted_organism == 'plants'"))
print("HIGH_mw_plant", HIGH_mw_plant)
HIGH_mw_fungi=len(extra.query("HIGH_mw==True and predicted_organism == 'fungi'"))
print("HIGH_mw_fungi", HIGH_mw_fungi)
HIGH_mw_bacteria=len(extra.query("HIGH_mw==True and predicted_organism == 'bacteria'"))
print("HIGH_mw_bacteria", HIGH_mw_bacteria)
print(HIGH_mw_plant+HIGH_mw_fungi+HIGH_mw_bacteria)
print("******")

MED_mw=len(extra.query("MED_mw==True and LOW_mw==False"))
print("MED_mw", MED_mw)
MED_mw_plants=len(extra.query("MED_mw==True and LOW_mw==False and predicted_organism == 'plants'"))
print("MED_mw_plants", MED_mw_plants)
MED_mw_fungi=len(extra.query("MED_mw==True and LOW_mw==False and predicted_organism == 'fungi'"))
print("MED_mw_fungi", MED_mw_fungi)
MED_mw_bacteria=len(extra.query("MED_mw==True and LOW_mw==False and predicted_organism == 'bacteria'"))
print("MED_mw_bacteria", MED_mw_bacteria)
print(MED_mw_bacteria+MED_mw_fungi+MED_mw_plants)
print("******")

MED_mw=len(extra.query("MW <= 800 and MW >= 300"))
print("MED_mw", MED_mw)
MED_mw_plants=len(extra.query("MW <= 800 and MW >= 300 and predicted_organism == 'plants'"))
print("MED_mw_plants", MED_mw_plants)
MED_mw_fungi=len(extra.query("MW <= 800 and MW >= 300 and predicted_organism == 'fungi'"))
print("MED_mw_fungi", MED_mw_fungi)
MED_mw_bacteria=len(extra.query("MW <= 800 and MW >= 300 and predicted_organism == 'bacteria'"))
print("MED_mw_bacteria", MED_mw_bacteria)
print(MED_mw_bacteria+MED_mw_fungi+MED_mw_plants)
print("******")



peptide=len(extra.query("isPeptide==True"))
print("peptide", peptide)
peptide_plant=len(extra.query("isPeptide==True and predicted_organism == 'plants'"))
print("peptide_plant", peptide_plant)
peptide_fungi=len(extra.query("isPeptide==True and predicted_organism == 'fungi'"))
print("peptide_fungi", peptide_fungi)
peptide_bacteria=len(extra.query("isPeptide==True and predicted_organism == 'bacteria'"))
print("peptide_bacteria", peptide_bacteria)
print("******")

sugar=len(extra.query("hasSugar==True"))
print("sugar", sugar)
sugar_plants=len(extra.query("hasSugar==True and origin ==0"))
print("sugar_plants", sugar_plants)
sugar_fungi=len(extra.query("hasSugar==True and predicted_organism == 'fungi'"))
print("sugar_fungi", sugar_fungi)
sugar_bacteria=len(extra.query("hasSugar==True and predicted_organism == 'bacteria'"))
print("sugar_bacteria", sugar_bacteria)
print("******")

plant = len(extra.query("predicted_organism == 'plants'"))
print("plant", plant)
fungi = len(extra.query("predicted_organism == 'fungi'"))
print("fungi", fungi)
bacteria = len(extra.query("predicted_organism == 'bacteria'"))
print("bacteria", bacteria)
print("******")
